reduced code duplication + other small fixes
