+++ lib/screen/profile_screen.dart
@@ -1,10 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
-
-import '../widgets/my_navigation_bar.dart';
 import '../widgets/outlined_button.dart';
 import '../widgets/playlist_card.dart';
-import '../widgets/tab_bar.dart';
+import '../widgets/profile_widgets.dart';
 
 class ProfileScreen extends StatefulWidget {
   const ProfileScreen({Key? key}) : super(key: key);
@@ -16,8 +14,6 @@ class ProfileScreen extends StatefulWidget {
 class _ProfileScreenState extends State<ProfileScreen>
     with TickerProviderStateMixin {
   late TabController _tabController;
-  
-  int numSongs = 15;
 
   @override
   void initState() {
@@ -37,7 +33,6 @@ class _ProfileScreenState extends State<ProfileScreen>
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      // bottomNavigationBar: const MyNavigationBar(),
       backgroundColor: Theme.of(context).colorScheme.background,
       body: Stack(children: [
         SvgPicture.asset(
@@ -54,35 +49,7 @@ class _ProfileScreenState extends State<ProfileScreen>
           child: Padding(
             padding: const EdgeInsets.fromLTRB(20, 85, 20, 20),
             child: Column(children: [
-              CircleAvatar(
-                foregroundImage: const NetworkImage(
-                  "https://picsum.photos/250?image=9",
-                ),
-                radius: 45,
-                child: Container(
-                  decoration: BoxDecoration(
-                    shape: BoxShape.circle,
-                    boxShadow: [
-                      BoxShadow(
-                        color: Theme.of(context).colorScheme.tertiary,
-                        blurRadius: 40,
-                        blurStyle: BlurStyle.outer,
-                      ),
-                    ],
-                  ),
-                ),
-              ),
-              const SizedBox(
-                height: 25,
-              ),
-              Text(
-                "Will Rice College",
-                textAlign: TextAlign.center,
-                style: Theme.of(context).textTheme.titleMedium,
-              ),
-              const SizedBox(
-                height: 25,
-              ),
+              const ProfileHeader(),
               Row(
                 mainAxisAlignment: MainAxisAlignment.center,
                 crossAxisAlignment: CrossAxisAlignment.center,
@@ -103,54 +70,11 @@ class _ProfileScreenState extends State<ProfileScreen>
                 ],
               ),
               const SizedBox(height: 10),
-              Row(
-                mainAxisAlignment: MainAxisAlignment.spaceAround,
-                crossAxisAlignment: CrossAxisAlignment.center,
-                children: [
-                  Column(
-                    children: [
-                      Text("208",
-                          style: Theme.of(context)
-                              .textTheme
-                              .displaySmall
-                              ?.copyWith(
-                                fontWeight: FontWeight.w600,
-                              )),
-                      Text("Followers",
-                          style: Theme.of(context).textTheme.displaySmall),
-                    ],
-                  ),
-                  Column(
-                    children: [
-                      Text("80",
-                          style: Theme.of(context)
-                              .textTheme
-                              .displaySmall
-                              ?.copyWith(fontWeight: FontWeight.w600)),
-                      Text("Following",
-                          style: Theme.of(context).textTheme.displaySmall),
-                    ],
-                  ),
-                ],
-              ),
+              const FollowCounts(followers: 200),
               const SizedBox(height: 10),
-              TabBars(tabController: _tabController,),
+              TabBars(tabController: _tabController, isPersonal: false),
               const SizedBox(height: 25),
               <Widget>[
-                Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    Label(name: "Songs", count: numSongs.toString()),
-                    PlaylistCard(),
-                    PlaylistCard(),
-                    PlaylistCard(),
-                    PlaylistCard(),
-                    PlaylistCard(),
-                    PlaylistCard(),
-                    PlaylistCard(),
-                    PlaylistCard(),
-                  ],
-                ),
                 Column(
                   crossAxisAlignment: CrossAxisAlignment.start,
                   children: const [
